✨ feat(scripts): astro project created
