+++ package.json
@@ -2,7 +2,7 @@
   "name": "templ",
   "type": "module",
   "private": true,
-  "packageManager": "pnpm@8.14.1",
+  "packageManager": "pnpm@8.14.2",
   "description": "A modern monorepo project template",
   "author": "Joydip Roy <joydipand@gmail.com> (https://github.com/rjoydip)",
   "license": "MIT",
@@ -46,7 +46,7 @@
     "better-commits": "^1.8.1",
     "eslint": "^8.56.0",
     "knip": "^3.13.2",
-    "pnpm": "^8.14.1",
+    "pnpm": "^8.14.2",
     "taze": "^0.13.1",
     "turbo": "^1.11.3"
   },
