+++ packages/config/test/index.test.ts
@@ -1,8 +1,8 @@
-import { join } from 'node:path'
+import { resolve } from 'node:path'
 import { describe, expect, it } from 'vitest'
 import { getBuildConfig } from '../src'
 
-const fixture = (folder: string) => join(join(__dirname, 'fixtures'), folder)
+const fixture = (folder: string) => resolve(__dirname, 'fixtures', folder)
 
 describe('@templ/config', () => {
   it('should be validate build data', async () => {
