+++ scripts/package.json
@@ -26,7 +26,9 @@
     "ci:env-copy": "tsx ./src/ci/env-copy.ts",
     "ci:env-delete": "tsx ./src/ci/env-delete.ts",
     "dist-check": "tsx ./src/dist-check.ts",
-    "create:stack": "tsx ./src/create/stack.ts",
+    "create:app": "tsx ./src/create/app.ts",
+    "create:doc": "tsx ./src/create/doc.ts",
+    "create:pkg": "tsx ./src/create/pkg.ts",
     "lint:size-limit": "tsx ./src/lint/size-limit.ts",
     "lint:size-report": "tsx ./src/lint/size-report.ts",
     "lint:type-cov": "tsx ./src/lint/type-cov.ts"
