+++ scripts/src/clean/others.ts
@@ -1,5 +1,5 @@
 import { cwd, exit } from 'node:process'
-import { join } from 'node:path'
+import { resolve } from 'node:path'
 import consola from 'consola'
 import { colors } from 'consola/utils'
 import { deleteAsync } from 'del'
@@ -10,13 +10,13 @@ async function main() {
 
   deletedPaths = await deleteAsync(['**/dist/**', '**/temp/**', '**/coverage/**'], {
     ignore: ignorePatterns,
-    cwd: join(cwd(), '..'),
+    cwd: resolve(cwd(), '..'),
     dryRun: hasDryRun(),
     force: true,
     absolute: false,
   })
 
-  deletedPaths.length ? consola.box(`Deleted files and directories:\n\n${deletedPaths.map(d => colors.magenta(d)).join('\n')}`) : consola.info('Nothing has been deleted')
+  deletedPaths.length ? consola.box(`Deleted files and directories:\n\n${deletedPaths.map(d => colors.magenta(d)).resolve('\n')}`) : consola.info('Nothing has been deleted')
 }
 
-main().catch(consola.error).finally(() => exit(0))
+main().catch(consola.error)
