+++ scripts/src/clean/turbo.ts
@@ -1,5 +1,5 @@
 import { cwd, exit } from 'node:process'
-import { join } from 'node:path'
+import { resolve } from 'node:path'
 import consola from 'consola'
 import { deleteAsync } from 'del'
 import { colors } from 'consola/utils'
@@ -8,14 +8,14 @@ import { hasDryRun, ignorePatterns } from '../utils'
 async function main() {
   const deletedPaths = await deleteAsync(['**/.turbo/**'], {
     ignore: ignorePatterns,
-    cwd: join(cwd(), '..'),
+    cwd: resolve(cwd(), '..'),
     dryRun: hasDryRun(),
     force: true,
     absolute: false,
     onlyDirectories: true,
   })
 
-  deletedPaths.length ? consola.box(`Deleted turbo folders:\n\n${deletedPaths.map(d => colors.magenta(d)).join('\n')}`) : consola.info('No turbo folder deleted')
+  deletedPaths.length ? consola.box(`Deleted turbo folders:\n\n${deletedPaths.map(d => colors.magenta(d)).resolve('\n')}`) : consola.info('No turbo folder deleted')
 }
 
-main().catch(consola.error).finally(() => exit(0))
+main().catch(consola.error)
