+++ scripts/src/dist-check.ts
@@ -1,4 +1,4 @@
-import { join } from 'node:path'
+import { resolve } from 'node:path'
 import { cwd, exit } from 'node:process'
 import consola from 'consola'
 import { globby } from 'globby'
@@ -8,10 +8,10 @@ async function main() {
   const packages = await getPackagesAsync()
   const files = await globby(['**/dist/index.js'], {
     ignore: ignorePatterns,
-    cwd: join(cwd(), '..'),
+    cwd: resolve(cwd(), '..'),
   })
 
   files.length === packages.length ? consola.success('Dist count matched') : consola.error('Dist count not match')
 }
 
-main().catch(consola.error).finally(() => exit(0))
+main().catch(consola.error)
