+++ templates/basic-ts/package.json
@@ -3,7 +3,7 @@
   "type": "module",
   "version": "0.0.0",
   "private": true,
-  "packageManager": "pnpm@8.14.1",
+  "packageManager": "pnpm@8.14.2",
   "description": "TEMPL basic-ts service",
   "license": "MIT",
   "homepage": "https://github.com/rjoydip/templ/tree/main/packages/basic-ts#readme",
